requirements.txt
PyMuPDF==1.22.3
pytesseract
pdf2image
Pillow
